--- lib/product_detal/widget/product_quantityselector.dart
@@ -1,179 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:shoppingmall/product_cart/product_cart.dart';
-
-class ProductQuantitySelector extends StatefulWidget {
-  final Map<String, dynamic> item;
-  final Function(int quantity)? onQuantityChanged;
-  final VoidCallback? onAddToCart;
-
-  const ProductQuantitySelector({
-    super.key,
-    required this.item,
-    this.onQuantityChanged,
-    this.onAddToCart,
-  });
-
-  @override
-  State<ProductQuantitySelector> createState() =>
-      _ProductQuantitySelectorState();
-}
-
-class _ProductQuantitySelectorState extends State<ProductQuantitySelector> {
-  int _quantity = 1;
-
-  void _incrementQuantity() {
-    setState(() {
-      _quantity++;
-    });
-    widget.onQuantityChanged?.call(_quantity);
-  }
-
-  void _decrementQuantity() {
-    if (_quantity > 1) {
-      setState(() {
-        _quantity--;
-      });
-      widget.onQuantityChanged?.call(_quantity);
-    }
-  }
-
-  void _goToCart(BuildContext context) {
-    Navigator.push(
-      context,
-      MaterialPageRoute(
-        builder: (context) => const ProductCart(),
-      ),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.all(16),
-      decoration: BoxDecoration(
-        color: Colors.white,
-        boxShadow: [
-          BoxShadow(
-            color: Colors.grey.withOpacity(0.2),
-            spreadRadius: 1,
-            blurRadius: 6,
-            offset: const Offset(0, -2),
-          ),
-        ],
-      ),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          // Price and quantity selector row
-          Container(
-            padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-            decoration: BoxDecoration(
-              color: Colors.grey[100],
-              borderRadius: BorderRadius.circular(8),
-            ),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                Container(
-                  padding:
-                      const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                  decoration: BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.circular(4),
-                  ),
-                  child: const Text(
-                    '3일이상소요',
-                    style: TextStyle(fontSize: 12),
-                  ),
-                ),
-                Text(
-                  '${widget.item['price']}원',
-                  style: const TextStyle(
-                    fontSize: 14,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              ],
-            ),
-          ),
-          const SizedBox(height: 16),
-          // Cart button and total price row
-          Row(
-            children: [
-              // Quantity selector
-              Container(
-                decoration: BoxDecoration(
-                  border: Border.all(color: Colors.grey[300]!),
-                  borderRadius: BorderRadius.circular(4),
-                ),
-                child: Row(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    IconButton(
-                      icon: const Icon(Icons.remove, size: 16),
-                      onPressed: _decrementQuantity,
-                      constraints: const BoxConstraints(
-                        minWidth: 32,
-                        minHeight: 32,
-                      ),
-                    ),
-                    Container(
-                      width: 40,
-                      alignment: Alignment.center,
-                      child: Text(
-                        _quantity.toString(),
-                        style: const TextStyle(fontSize: 14),
-                      ),
-                    ),
-                    IconButton(
-                      icon: const Icon(Icons.add, size: 16),
-                      onPressed: _incrementQuantity,
-                      constraints: const BoxConstraints(
-                        minWidth: 32,
-                        minHeight: 32,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              const Spacer(),
-              // Total price
-              Text(
-                '${(widget.item['price'] * _quantity).toString()}원',
-                style: const TextStyle(
-                  fontSize: 16,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-              const SizedBox(width: 12),
-              // Add to cart button
-              ElevatedButton(
-                onPressed: () {
-                  widget.onAddToCart?.call(); // 기존 콜백 호출
-                  _goToCart(context); // ProductCart 페이지로 이동
-                },
-                style: ElevatedButton.styleFrom(
-                  backgroundColor: Colors.blue,
-                  padding: const EdgeInsets.symmetric(
-                    horizontal: 24,
-                    vertical: 12,
-                  ),
-                  shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(8),
-                  ),
-                ),
-                child: const Text(
-                  '장바구니로',
-                  style: TextStyle(
-                    fontSize: 14,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-}
